R with application in simple Statistics
R is a powerful programming language widely used for simple statistical analyses. It offers a range of functions and packages that facilitate tasks such as calculating descriptive statistics, conducting hypothesis tests, performing regression analyses, and generating visualizations. With its intuitive syntax and extensive library of statistical tools, R provides researchers and data analysts with a flexible and efficient environment to explore and interpret data for making informed decisions. Its popularity and active community support make it an ideal choice for beginners and experts alike in the field of statistics.